+++ haskell/app/Trader/Http.hs
@@ -14,7 +14,7 @@ module Trader.Http (
 ) where
 
 import Control.Concurrent (MVar, modifyMVar, newMVar, threadDelay)
-import Control.Exception (SomeException, displayException, throwIO, try)
+import Control.Exception (displayException, throwIO, try)
 import Data.ByteString.Char8 (ByteString)
 import qualified Data.ByteString.Char8 as BS
 import qualified Data.ByteString.Lazy as BL
@@ -115,7 +115,7 @@ httpLbsWithRetry cfg mLabel mgr req0 = go 0
     go attempt = do
         applyRateLimit hostBs
         t0 <- getTimeMs
-        respOrErr <- try (httpLbs req mgr) :: IO (Either SomeException (Response BL.ByteString))
+        respOrErr <- try (httpLbs req mgr) :: IO (Either HttpException (Response BL.ByteString))
         t1 <- getTimeMs
         let latencyMs = max 0 (fromIntegral (t1 - t0) :: Int)
         case respOrErr of
@@ -319,7 +319,7 @@ logHttpAttempt ::
     ByteString ->
     ByteString ->
     ByteString ->
-    Either SomeException (Response BL.ByteString) ->
+    Either HttpException (Response BL.ByteString) ->
     Int ->
     Int ->
     Bool ->